Fixed labjack header file to return float (not uint16_t), included blast.h in cryostat.c so that we can use blast formated print statements.

+++ mcp/include/labjack.h
@@ -39,7 +39,7 @@
 #define LJ_R_FPA_1K_IND 2
 
 void labjack_networking_init(int m_which, size_t m_numchannels, size_t m_scans_per_packet);
-uint16_t labjack_get_value(int m_labjack, int m_channel);
+float labjack_get_value(int m_labjack, int m_channel);
 void initialize_labjack_commands(int m_which);
 void store_labjack_data(void);
